% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peg_get.R
\name{peg_get}
\alias{peg_get}
\title{Get a dataset from the Winnipeg Open Data Portal}
\usage{
peg_get(dataset_id)
}
\arguments{
\item{dataset_id}{A character string of the dataset ID e.g. \code{"d4mq-wa44"}}
}
\value{
A tibble of the dataset
}
\description{
Get a dataset from the Winnipeg Open Data Portal
}
\examples{
\dontrun{
peg_get("d4mq-wa44")
}
}
