% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peg_info.R
\name{peg_info}
\alias{peg_info}
\title{Get dataset-level information from the Winnipeg Open Data Portal}
\usage{
peg_info(dataset_id)
}
\arguments{
\item{dataset_id}{A character string of the dataset ID e.g. \code{"d4mq-wa44"}}
}
\value{
A tibble with one row containing dataset-level metadata including
name, description, category, timestamps, counts, tags, license,
and provenance.
}
\description{
Get dataset-level information from the Winnipeg Open Data Portal
}
\examples{
\dontrun{
peg_info("d4mq-wa44")
}
}
